+++ src/tad/WPBrowser/Environment/Configuration.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Models the configuration for a test WordPress installation like the one set up by WPLoader.
+ *
+ * @package tad\WPBrowser\Configuration
+ */
+
+namespace tad\WPBrowser\Configuration;
+
+use tad\WPBrowser\Utils\Map;
+
+/**
+ * Class TestInstallationConfiguration
+ *
+ * @package tad\WPBrowser\Configuration
+ */
+class Configuration extends Map
+{
+}