+++ src/tad/WPBrowser/Utils/Configuration.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Models the configuration for a test WordPress installation like the one set up by WPLoader.
+ *
+ * @package tad\WPBrowser\Utils
+ */
+
+namespace tad\WPBrowser\Utils;
+
+/**
+ * Class TestInstallationConfiguration
+ *
+ * @package tad\WPBrowser\Utils
+ */
+class Configuration extends Map
+{
+}