+++ src/tad/WPBrowser/utils.php
@@ -19,3 +19,4 @@ require_once __DIR__ . '/patchwork.php';
 require_once __DIR__ . '/filesystem.php';
 require_once __DIR__ . '/events.php';
 require_once __DIR__ . '/deprecated.php';
+require_once __DIR__ . '/wp.php';